lms/bot.py
from telegram.ext import Updater, CommandHandler, MessageHandler, Filters, CallbackContext
from telegram import Update


def start(update: Update, context: CallbackContext):
    update.message.reply_text('Hello!')

